mds: fix mdsmap->get_metadata_pool() return to int64_t

+++ src/tools/cephfs/Resetter.cc
@@ -34,7 +34,7 @@ int Resetter::reset(mds_role_t role)
 
   auto fs =  fsmap->get_filesystem(role.fscid);
   assert(fs != nullptr);
-  int const pool_id = fs->mds_map.get_metadata_pool();
+  int64_t const pool_id = fs->mds_map.get_metadata_pool();
 
   JournalPointer jp(role.rank, pool_id);
   int jp_load_result = jp.load(objecter);
@@ -115,7 +115,7 @@ int Resetter::reset_hard(mds_role_t role)
 {
   auto fs =  fsmap->get_filesystem(role.fscid);
   assert(fs != nullptr);
-  int const pool_id = fs->mds_map.get_metadata_pool();
+  int64_t const pool_id = fs->mds_map.get_metadata_pool();
 
   JournalPointer jp(role.rank, pool_id);
   jp.front = role.rank + MDS_INO_LOG_OFFSET;
